+++ frigate/types.py
@@ -24,6 +24,10 @@ class CameraMetricsTypes(TypedDict):
     skipped_fps: Synchronized
 
 
+class RecordMetricsTypes(TypedDict):
+    record_enabled: Synchronized
+
+
 class StatsTrackingTypes(TypedDict):
     camera_metrics: dict[str, CameraMetricsTypes]
     detectors: dict[str, ObjectDetectProcess]
